Add example, clean up readme, alignments
