+++ src/layout/linear/mod.rs
@@ -1,230 +1,20 @@
 //! Linear layout
 
-use super::*;
-use crate::align::{Alignment, HorizontalAlignment, VerticalAlignment};
+use crate::{
+    align::{HorizontalAlignment, VerticalAlignment},
+    layout::{ChainTerminator, ViewChainElement, ViewGroup, ViewLink},
+    prelude::*,
+};
+use embedded_graphics::primitives::Rectangle;
 
-/// Secondary alignment is used to align views perpendicular to the placement axis.
-///
-/// For example, use `horizontal::Right` to align views to the right in a vertical linear layout.
-pub trait SecondaryAlignment: Alignment {
-    fn measure(prev: Size, view: &impl View) -> Size;
-}
+mod layout_direction;
+mod layout_operation;
+mod secondary_alignment;
 
-fn max_width(prev_size: Size, view: &impl View) -> Size {
-    let view_size = RectExt::size(&view.bounds());
+pub use layout_direction::{Horizontal, LayoutDirection, Vertical};
+pub use secondary_alignment::SecondaryAlignment;
 
-    Size::new(
-        prev_size.width.max(view_size.width),
-        prev_size.height + view_size.height,
-    )
-}
-
-fn cascading(prev_size: Size, view: &impl View) -> Size {
-    let view_size = RectExt::size(&view.bounds());
-
-    Size::new(
-        prev_size.width + view_size.width,
-        prev_size.height + view_size.height,
-    )
-}
-
-impl SecondaryAlignment for horizontal::Left {
-    fn measure(prev_size: Size, view: &impl View) -> Size {
-        max_width(prev_size, view)
-    }
-}
-impl SecondaryAlignment for horizontal::Center {
-    fn measure(prev_size: Size, view: &impl View) -> Size {
-        max_width(prev_size, view)
-    }
-}
-impl SecondaryAlignment for horizontal::Right {
-    fn measure(prev_size: Size, view: &impl View) -> Size {
-        max_width(prev_size, view)
-    }
-}
-impl SecondaryAlignment for horizontal::RightToLeft {
-    fn measure(prev_size: Size, view: &impl View) -> Size {
-        cascading(prev_size, view)
-    }
-}
-impl SecondaryAlignment for horizontal::LeftToRight {
-    fn measure(prev_size: Size, view: &impl View) -> Size {
-        cascading(prev_size, view)
-    }
-}
-
-fn max_height(prev_size: Size, view: &impl View) -> Size {
-    let view_size = RectExt::size(&view.bounds());
-
-    Size::new(
-        prev_size.width + view_size.width,
-        prev_size.height.max(view_size.height),
-    )
-}
-
-impl SecondaryAlignment for vertical::Top {
-    fn measure(prev_size: Size, view: &impl View) -> Size {
-        max_height(prev_size, view)
-    }
-}
-impl SecondaryAlignment for vertical::Center {
-    fn measure(prev_size: Size, view: &impl View) -> Size {
-        max_height(prev_size, view)
-    }
-}
-impl SecondaryAlignment for vertical::Bottom {
-    fn measure(prev_size: Size, view: &impl View) -> Size {
-        max_height(prev_size, view)
-    }
-}
-impl SecondaryAlignment for vertical::TopToBottom {
-    fn measure(prev_size: Size, view: &impl View) -> Size {
-        cascading(prev_size, view)
-    }
-}
-impl SecondaryAlignment for vertical::BottomToTop {
-    fn measure(prev_size: Size, view: &impl View) -> Size {
-        cascading(prev_size, view)
-    }
-}
-
-/// Helper trait that describes a layout direction.
-pub trait LayoutDirection: Copy + Clone {}
-pub trait LayoutOperation<LD: LayoutDirection> {
-    fn measure(&self) -> Size;
-    fn arrange(&mut self, bounds: Rectangle) -> Rectangle;
-}
-
-/// Horizontal layout direction
-#[derive(Copy, Clone)]
-pub struct Horizontal<Secondary: SecondaryAlignment + VerticalAlignment> {
-    secondary: Secondary,
-}
-
-impl Default for Horizontal<vertical::Bottom> {
-    fn default() -> Self {
-        Self {
-            secondary: vertical::Bottom,
-        }
-    }
-}
-
-impl<Secondary> LayoutDirection for Horizontal<Secondary> where
-    Secondary: SecondaryAlignment + VerticalAlignment
-{
-}
-
-impl<V, VCE, Secondary> LayoutOperation<Horizontal<Secondary>> for ViewLink<V, VCE>
-where
-    V: View + Align,
-    VCE: ViewChainElement + LayoutOperation<Horizontal<Secondary>>,
-    Secondary: SecondaryAlignment + VerticalAlignment,
-{
-    fn measure(&self) -> Size {
-        // Counting this way assumes that views are aligned and not cascading in the other direction
-        let current_el_size = RectExt::size(&self.view.bounds());
-        if VCE::IS_TERMINATOR {
-            current_el_size
-        } else {
-            let prev_size = self.next.measure();
-            Secondary::measure(prev_size, &self.view)
-        }
-    }
-
-    fn arrange(&mut self, bounds: Rectangle) -> Rectangle {
-        if VCE::IS_TERMINATOR {
-            self.view
-                .align_to(&bounds, horizontal::Left, Secondary::new());
-        } else {
-            let previous = self.next.arrange(bounds);
-
-            self.view.align_to(
-                &previous.bounds(),
-                horizontal::LeftToRight,
-                Secondary::new(),
-            );
-        }
-        self.view.bounds()
-    }
-}
-
-impl<Secondary> LayoutOperation<Horizontal<Secondary>> for ChainTerminator
-where
-    Secondary: SecondaryAlignment + VerticalAlignment,
-{
-    fn measure(&self) -> Size {
-        Size::new(0, 0)
-    }
-
-    fn arrange(&mut self, _bounds: Rectangle) -> Rectangle {
-        // Nothing to do
-        Rectangle::new(Point::zero(), Point::zero())
-    }
-}
-
-/// Vertical layout direction
-#[derive(Copy, Clone)]
-pub struct Vertical<Secondary: SecondaryAlignment + HorizontalAlignment> {
-    secondary: Secondary,
-}
-
-impl Default for Vertical<horizontal::Left> {
-    fn default() -> Self {
-        Self {
-            secondary: horizontal::Left,
-        }
-    }
-}
-
-impl<Secondary> LayoutDirection for Vertical<Secondary> where
-    Secondary: SecondaryAlignment + HorizontalAlignment
-{
-}
-
-impl<V, VCE, Secondary> LayoutOperation<Vertical<Secondary>> for ViewLink<V, VCE>
-where
-    V: View + Align,
-    VCE: ViewChainElement + LayoutOperation<Vertical<Secondary>>,
-    Secondary: SecondaryAlignment + HorizontalAlignment,
-{
-    fn measure(&self) -> Size {
-        // Counting this way assumes that views are aligned and not cascading in the other direction
-        let current_el_size = RectExt::size(&self.view.bounds());
-        if VCE::IS_TERMINATOR {
-            current_el_size
-        } else {
-            let prev_size = self.next.measure();
-            Secondary::measure(prev_size, &self.view)
-        }
-    }
-
-    fn arrange(&mut self, bounds: Rectangle) -> Rectangle {
-        if VCE::IS_TERMINATOR {
-            self.view.align_to(&bounds, Secondary::new(), vertical::Top);
-        } else {
-            let previous = self.next.arrange(bounds);
-
-            self.view
-                .align_to(&previous.bounds(), Secondary::new(), vertical::TopToBottom);
-        }
-        self.view.bounds()
-    }
-}
-
-impl<Secondary> LayoutOperation<Vertical<Secondary>> for ChainTerminator
-where
-    Secondary: SecondaryAlignment + HorizontalAlignment,
-{
-    fn measure(&self) -> Size {
-        Size::new(0, 0)
-    }
-
-    fn arrange(&mut self, _bounds: Rectangle) -> Rectangle {
-        // Nothing to do
-        Rectangle::new(Point::zero(), Point::zero())
-    }
-}
+use layout_operation::LayoutOperation;
 
 /// LinearLayout
 ///
